src: split profiling helpers from util
