+++ src/node_v8.cc
@@ -26,6 +26,7 @@
 #include "memory_tracker-inl.h"
 #include "node.h"
 #include "node_external_reference.h"
+#include "node_profiling.h"
 #include "util-inl.h"
 #include "v8-profiler.h"
 #include "v8.h"
